Move api docs internal tests to a separate package
