+++ pkg/agent/api_docs_test.go
@@ -1,7 +1,8 @@
-package agent
+package agent_test
 
 import (
 	"context"
+	"github.com/emissary-ingress/emissary/v3/pkg/agent"
 	"net/url"
 	"testing"
 
@@ -27,7 +28,7 @@ func TestAPIDocsStore(t *testing.T) {
 
 		expectedRequestURL     string
 		expectedRequestHost    string
-		expectedRequestHeaders []Header
+		expectedRequestHeaders []agent.Header
 		expectedSOTW           []*snapshotTypes.APIDoc
 	}
 	cases := []*testCases{
@@ -95,7 +96,7 @@ func TestAPIDocsStore(t *testing.T) {
 
 			expectedRequestURL:     "https://some-svc.fqdn:443/internal-path/docs-location",
 			expectedRequestHost:    "",
-			expectedRequestHeaders: []Header{{Name: "header-key", Value: "header-value"}},
+			expectedRequestHeaders: []agent.Header{{Name: "header-key", Value: "header-value"}},
 			expectedSOTW:           []*snapshotTypes.APIDoc{},
 		},
 		{
@@ -126,7 +127,7 @@ func TestAPIDocsStore(t *testing.T) {
 
 			expectedRequestURL:     "http://some-svc.default:8080/docs-location",
 			expectedRequestHost:    "mapping-hostname",
-			expectedRequestHeaders: []Header{},
+			expectedRequestHeaders: []agent.Header{},
 			expectedSOTW: []*snapshotTypes.APIDoc{{
 				TypeMeta: &kates.TypeMeta{
 					Kind:       "OpenAPI",
@@ -169,7 +170,7 @@ func TestAPIDocsStore(t *testing.T) {
 
 			expectedRequestURL:     "https://external-url",
 			expectedRequestHost:    "",
-			expectedRequestHeaders: []Header{},
+			expectedRequestHeaders: []agent.Header{},
 			expectedSOTW: []*snapshotTypes.APIDoc{{
 				TypeMeta: &kates.TypeMeta{
 					Kind:       "OpenAPI",
@@ -201,7 +202,7 @@ func TestAPIDocsStore(t *testing.T) {
 				},
 			}
 
-			store := NewAPIDocsStore()
+			store := agent.NewAPIDocsStore()
 			store.Client = NewMockAPIDocsHTTPClient(t, c.expectedRequestURL, c.expectedRequestHost, c.expectedRequestHeaders, c.rawJSONDocsContent, c.JSONDocsErr)
 
 			// Processing the test case snapshot should yield the expected state of the world
@@ -215,121 +216,7 @@ func TestAPIDocsStore(t *testing.T) {
 	}
 }
 
-func TestParseToOpenAPIV3ConvertToV2(t *testing.T) {
-	// given
-	doc := `{
-    "basePath": "/api",
-    "definitions": {
-        "Account": {
-            "properties": {
-                "avatarUrl": {
-                    "description": "Avatar url",
-                    "type": "string"
-                },
-                "id": {
-                    "description": "Account id",
-                    "type": "string"
-                },
-                "identityProvider": {
-                    "description": "Identity Provider",
-                    "type": "string"
-                },
-                "invitationPending": {
-                    "description": "The user has a pending invitation for this account",
-                    "type": "boolean"
-                },
-                "name": {
-                    "description": "Account name",
-                    "type": "string"
-                }
-            },
-            "required": [
-                "id",
-                "name"
-            ],
-            "type": "object"
-        },
-        "Error": {
-            "properties": {
-                "errorCode": {
-                    "enum": [
-                        "DEMO_CLUSTER_DEPLETED",
-                        "SUBSCRIPTION_LIMIT_REACHED"
-                    ],
-                    "type": "string",
-                    "x-nullable": true
-                },
-                "message": {
-                    "type": "string"
-                }
-            },
-            "required": [
-                "message"
-            ],
-            "type": "object"
-        }
-    },
-    "paths": {
-        "/accounts": {
-            "get": {
-                "operationId": "listAccounts",
-                "produces": [
-                    "application/json"
-                ],
-                "responses": {
-                    "200": {
-                        "description": "successful operation",
-                        "schema": {
-                            "items": {
-                                "$ref": "#/definitions/Account"
-                            },
-                            "type": "array"
-                        }
-                    },
-                    "500": {
-                        "$ref": "#/responses/GenericError"
-                    }
-                },
-                "summary": "List the accounts the current user is a member of",
-                "tags": [
-                    "account"
-                ]
-            }
-        }
-    },
-    "responses": {
-        "GenericError": {
-            "description": "An unexpected occurred on the server",
-            "schema": {
-                "$ref": "#/definitions/Error"
-            }
-        }
-    },
-    "schemes": [
-        "http"
-    ],
-    "servers": [
-        {
-            "url": "https://23.251.148.46/backend"
-        }
-    ],
-    "swagger": "2.0",
-    "tags": [
-        {
-            "description": "The Account API",
-            "name": "account"
-        }
-    ]
-}`
-	// when
-	v3Doc, err := parseToOpenAPIV3([]byte(doc))
-
-	// then
-	assert.NoError(t, err)
-	assert.NotNil(t, v3Doc)
-}
-
-func NewMockAPIDocsHTTPClient(t *testing.T, expectedRequestURL string, expectedRequestHost string, expectedRequestHeaders []Header, content string, err error) APIDocsHTTPClient {
+func NewMockAPIDocsHTTPClient(t *testing.T, expectedRequestURL string, expectedRequestHost string, expectedRequestHeaders []agent.Header, content string, err error) agent.APIDocsHTTPClient {
 	return &mockAPIDocsHTTPClient{
 		t:                      t,
 		expectedRequestURL:     expectedRequestURL,
@@ -345,13 +232,13 @@ type mockAPIDocsHTTPClient struct {
 
 	expectedRequestURL     string
 	expectedRequestHost    string
-	expectedRequestHeaders []Header
+	expectedRequestHeaders []agent.Header
 
 	resultContent string
 	resultErr     error
 }
 
-func (c *mockAPIDocsHTTPClient) Get(ctx context.Context, requestURL *url.URL, requestHost string, requestHeaders []Header) ([]byte, error) {
+func (c *mockAPIDocsHTTPClient) Get(ctx context.Context, requestURL *url.URL, requestHost string, requestHeaders []agent.Header) ([]byte, error) {
 	if c.expectedRequestURL == "" {
 		c.t.Errorf("unexpected call to APIDocsHTTPClient.Get")
 		c.t.Fail()
